+++ pkg/cluster/controller/projects/approvalrules/zz_controller.go
@@ -236,6 +236,6 @@ func (e *external) Disconnect(ctx context.Context) error {
 }
 
 func (e *external) updateExternalName(ctx context.Context, cr *v1alpha1.ApprovalRule, approvalRule *gitlab.ProjectApprovalRule) error {
-	meta.SetExternalName(cr, strconv.FormatInt(int64(approvalRule.ID), 10))
+	meta.SetExternalName(cr, strconv.FormatInt(approvalRule.ID, 10))
 	return e.kube.Update(ctx, cr)
 }
