+++ pkg/cluster/controller/projects/runners/zz_controller.go
@@ -190,7 +190,7 @@ func (e *external) Create(ctx context.Context, mg resource.Managed) (managed.Ext
 		return managed.ExternalCreation{}, errors.Wrap(err, errCreateFailed)
 	}
 
-	meta.SetExternalName(cr, strconv.FormatInt(int64(runner.ID), 10))
+	meta.SetExternalName(cr, strconv.FormatInt(runner.ID, 10))
 
 	if runner.TokenExpiresAt != nil {
 		t := metav1.NewTime(*runner.TokenExpiresAt)
