+++ pkg/namespaced/clients/projects/project_test.go
@@ -293,7 +293,7 @@ func TestGenerateObservation(t *testing.T) {
 						RepositorySize:   storageStatistics.RepositorySize,
 						LFSObjectsSize:   storageStatistics.LFSObjectsSize,
 						JobArtifactsSize: storageStatistics.JobArtifactsSize,
-						CommitCount:      int64(projectStatisticsCommitCount),
+						CommitCount:      projectStatisticsCommitCount,
 					},
 					Links: &gitlab.Links{
 						Self: linksSelf,
